PSMDB-712: Add connection pool support to LDAP auth
Issue: the ldap query function used a single connection for all queries
with a global mutex. With referral support turned on searches could
take more time, resulting in decreasing login performance.

Fix: this commit changes the single connection ldap manager
implementation to a connection pool, which is able to handle multiple
login attemts at the same time.

To make this configurable, it introduduses a new ldap configuration
variable, "max_pool_size", which limits the number of active ldap
connections. When reaching this limit, the behavior falls back to the
previous one: the authentication attempt will wait until a connection is
available in the pool.